using correct taxon-id for genus matches

+++ jacq-common/src/main/java/org/jacq/common/model/names/taxamatch/Searchresult.java
@@ -12,7 +12,7 @@ public class Searchresult {
     private String taxon;
     private Long ID;
     private List<Species> species = new ArrayList<Species>();
-    private Integer taxonID;
+    private Long taxonID;
 
     /**
      *
@@ -112,11 +112,11 @@ public class Searchresult {
     }
 
     @XmlElement(name = "taxonID")
-    public Integer getTaxonID() {
+    public Long getTaxonID() {
         return taxonID;
     }
 
-    public void setTaxonID(Integer taxonID) {
+    public void setTaxonID(Long taxonID) {
         this.taxonID = taxonID;
     }
